requirements.txt
flask
gunicorn
python-dotenv
flask-limiter
flask-cors
pyOpenSSL
langchain-community
langchain-core
langchain-anthropic
langchain-google-genai
langchain-openai
langchain-text-splitters
langchain-pinecone
langchain
langsmith
langgraph
openai
pinecone
weave==0.51.31
PyYAML
beautifulsoup4